Made the Login system more efficient .
We got rid of the table for passwords and added them to the user table. This hopefully make the system a bit more secure and easier to read. After that, we also added a query to check the name of the role by checking the role_id - therefore, we are now ready to just copy paste respective dashboard paths into lines 53 and 57.